+++ src/client-agent/agentd.c
@@ -83,7 +83,7 @@ void AgentdStart(char *dir, int uid, int gid, char *user, char *group)
 
 
     /* Start up message */
-    verbose(STARTUP_MSG, ARGV0, getpid());
+    verbose(STARTUP_MSG, ARGV0, (int)getpid());
 
     
     /* Initial random numbers */
