+++ src/os_dbd/main.c
@@ -249,7 +249,7 @@ int main(int argc, char **argv)
 
     
     /* Start up message */
-    verbose(STARTUP_MSG, ARGV0, getpid());
+    verbose(STARTUP_MSG, ARGV0, (int)getpid());
     
 
     /* the real daemon now */
